+++ app/routes/scrapping/scrapping.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
-from services.scrapping.index import scrapping_test, scrap
+from services.scrapping.index import scrapping_test, scrap, single_scrap_carrefour
 
 scrapping_bp = Blueprint('scrapping', __name__)
 
@@ -10,3 +10,8 @@ def test():
 @scrapping_bp.route('/', methods=['POST'])
 def index():
     return scrap()
+
+
+@scrapping_bp.route('/carrefour', methods=['POST'])
+def carrefour():
+    return single_scrap_carrefour()