+++ app/services/db/seed/market.py
@@ -13,17 +13,17 @@ def seed_markets():
 
     cursor = connection.cursor()
 
-    #create table if not exists
+    #create table if not exists name should be unique
     cursor.execute(
         """
         CREATE TABLE IF NOT EXISTS markets (
             id SERIAL PRIMARY KEY,
-            name VARCHAR(255) NOT NULL
+            name VARCHAR(255) NOT NULL UNIQUE
         )
         """
     )
 
-    markets_to_seed = ["coto", "jumbo"]
+    markets_to_seed = ["coto", "jumbo", "carrefour"]
 
     for market in markets_to_seed:
         cursor.execute(